refactor: split ProjectService into domain-based files
